FEC: extract PacketDispatcher to separate file
